int to size_t - make the api more consistent
git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@163 1dbd42e5-0d20-4692-8756-183d5fc073b5

+++ modules/c++/math.linear/include/math/linear/VectorN.h
@@ -172,7 +172,7 @@ public:
 #endif
         return mRaw[i][0];
     }
-    inline _T& operator[](int i)
+    inline _T& operator[](size_t i)
     {
 #if defined(MATH_LINEAR_BOUNDS)
         assert( i < _ND );
